+++ app/Http/Controllers/Api/v1/ProductController.php
@@ -6,11 +6,13 @@ use App\DTO\ProductDTO;
 use App\Models\Product;
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
-use App\Services\contract\ProductServiceInterface;
 use App\Http\Controllers\Api\v1\BaseApiController;
+use App\Services\contract\ProductServiceInterface;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ProductController extends BaseApiController
 {
+    use AuthorizesRequests;
     public function __construct(public ProductServiceInterface $service)
     {
     }
@@ -26,6 +28,9 @@ class ProductController extends BaseApiController
 
     public function store(StoreProductRequest $request)
     {
+
+
+
         $DTO = ProductDTO::fromRequest($request);
         $product = $this->service->store($DTO);
         return $this->sendResponse(
@@ -58,6 +63,7 @@ class ProductController extends BaseApiController
 
     public function destroy(Product $product)
     {
+        $this->authorize("delete", $product);
         $this->service->delete($product);
 
         return $this->sendResponse(message: "product deleted", result: true, code: 204);
